Implemented local-time display using full timezone name (e.g., Eastern Standard Time) for webinar times. Updated helper to format with timeZoneName: "long" and integrated across webinar-related components.

X-Lovable-Edit-ID: edt-969cecf0-2d4f-48d6-91ac-c5128bfd7608

+++ src/lib/webinar.ts
@@ -8,6 +8,6 @@ export function getLocalWebinarTime(): string {
     day: "numeric",
     hour: "numeric",
     minute: "2-digit",
-    timeZoneName: "short",
+    timeZoneName: "long",
   });
 }
